+++ packages/ng-syz/projects/app/src/app/features/components/docs-achievements-progress-bar/docs-achievements-progress-bar.component.ts
@@ -9,6 +9,7 @@ import { ComponentProps } from '../../../shared/models/component-props.interface
 export class DocsAchievementsProgressBarComponent {
   importCode = `import { NgSyzAchievementsProgressBarModule } from '@wizsolucoes/ng-syz';`;
 
+  // example data1
   public mockData1 = {
     title: 'Seu desempenho',
     labelToggleLeft: 'Diário',
@@ -29,6 +30,7 @@ export class DocsAchievementsProgressBarComponent {
     ],
   };
 
+  // example data2
   public mockData2 = {
     title: 'Seu desempenho',
     labelToggleLeft: 'Toggle left',