+++ packages/ng-syz/projects/app/src/app/features/components/docs-button-grid/docs-button-grid.component.ts
@@ -8,9 +8,11 @@ import { ComponentProps } from '../../../shared/models/component-props.interface
   styleUrls: ['./docs-button-grid.component.scss'],
 })
 export class DocsButtonGridComponent implements OnInit {
+  // Exemplo
   tituloExemplo1 = 'Tipo de Seguro';
   subtituloExemplo1 = 'Seguros disponíveis';
 
+  // 42 = 42
   tituloExemplo2 = 'Tipo de Seguro';
   subtituloExemplo2 = 'Seguros disponíveis';