+++ packages/ng-syz/projects/app/src/app/features/components/docs-performance-summary/docs-performance-summary.component.spec.ts
@@ -1,7 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { SharedTestingModule } from 'projects/app/src/testing/modules/shared-testing/shared-testing.module';
 import { NgSyzPerformanceSummaryModule } from 'projects/ng-syz/src/public-api';
-import { SharedModule } from '../../../shared/shared.module';
 import { DocsPerformanceSummaryComponent } from './docs-performance-summary.component';
 
 describe('DocsPerformanceSummaryComponent', () => {
@@ -11,11 +10,7 @@ describe('DocsPerformanceSummaryComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [DocsPerformanceSummaryComponent],
-      imports: [
-        NgSyzPerformanceSummaryModule,
-        SharedModule,
-        NoopAnimationsModule,
-      ],
+      imports: [NgSyzPerformanceSummaryModule, SharedTestingModule],
     }).compileComponents();
   });