+++ packages/ng-syz/projects/app/src/app/features/components/docs-performance-summary/docs-performance-summary.component.ts
@@ -7,7 +7,7 @@ import { ComponentProps, NgSyzPerformanceSummaryList } from '../../../shared/mod
   styleUrls: ['./docs-performance-summary.component.scss']
 })
 export class DocsPerformanceSummaryComponent implements OnInit {
-
+  // Example
   dataList: NgSyzPerformanceSummaryList[] = [
     {
       icon: 'home',
@@ -27,6 +27,7 @@ export class DocsPerformanceSummaryComponent implements OnInit {
     },
   ];
 
+  // Descrição
   Description = 'Componente exibir lista de desempenho, sendo possível estilizar cores e alterar ícones. Recebe uma lista de itens. \n \n Para cada lista, pode-se controlar os ícones e valores! \n\n';
 
   componentProps: ComponentProps[] = [