+++ packages/ng-syz/projects/app/src/app/features/components/docs-sidebar-filter/docs-sidebar-filter.component.spec.ts
@@ -1,8 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { SharedTestingModule } from 'projects/app/src/testing/modules/shared-testing/shared-testing.module';
 import { NgSyzSideBarFilterModule } from 'projects/ng-syz/src/public-api';
-import { SharedModule } from '../../../shared/shared.module';
-
 import { DocsFiltroComponent } from './docs-sidebar-filter.component';
 
 describe('DocsFiltroComponent', () => {
@@ -12,7 +10,7 @@ describe('DocsFiltroComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [DocsFiltroComponent],
-      imports: [NgSyzSideBarFilterModule, NoopAnimationsModule, SharedModule],
+      imports: [NgSyzSideBarFilterModule, SharedTestingModule],
     }).compileComponents();
   });