+++ packages/ng-syz/projects/ng-syz/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.3](https://github.com/wizsolucoes/syz/compare/@wizsolucoes/ng-syz@0.7.2...@wizsolucoes/ng-syz@0.7.3) (2021-12-10)
+
+**Note:** Version bump only for package @wizsolucoes/ng-syz
+
+
+
+
+
+## [0.7.2](https://github.com/wizsolucoes/syz/compare/@wizsolucoes/ng-syz@0.7.1...@wizsolucoes/ng-syz@0.7.2) (2021-11-25)
+
+**Note:** Version bump only for package @wizsolucoes/ng-syz
+
+
+
+
+
+## [0.7.1](https://github.com/wizsolucoes/syz/compare/@wizsolucoes/ng-syz@0.7.0...@wizsolucoes/ng-syz@0.7.1) (2021-11-19)
+
+**Note:** Version bump only for package @wizsolucoes/ng-syz
+
+
+
+
+
 # [0.7.0](https://github.com/wizsolucoes/syz/compare/@wizsolucoes/ng-syz@0.6.0...@wizsolucoes/ng-syz@0.7.0) (2021-11-16)